// src/app/models/patient-model.ts
export class PatientModel {

  password: string;

  firstName: string;

  lastName: string;

  email: string;

  otherName: string;

  phoneNumber: string;

  phoneNumberObj: any;

  otherPhoneNumber: string;

  otherPhoneNumberObj: any;

  gender: string;

  dob: any;

  address: string;

  twoFactor: boolean;

  nextOFKinFirstName: string;

  nextOFKinLastName: string ;

  nextOFKinPhoneNumber: string;

  nextOFKinPhoneNumberObj: string;

}